an initial and failed attempt to try modules with bazel build
The attemt failed due to a sealed LLVM bazel module. The module implicitly adds a flag, disabling c++20 modules.

In my opiniion reproducable env is more significant than fast builing. So, I'm waiting for a moment when Buzzel start support modules out of the box... Do not be the evel, isn't it?